requirements.txt
gensim==4.1.2
mysql-connector-python==8.0.23
pandas==1.3.3
requests==2.26.0
streamlit==1.10.0
